requirements.txt
requests==2.31.0
requests-cache==0.6.3
pillow==10.2.0
opencv-python-headless==4.7.0.72
fluent-logger==0.10.0
numpy==1.24.4
m3u8==3.4.0